pyproject.toml
[project]
name = "mcp-extendicare-ppt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.9.4",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "python-pptx>=1.0.2",
    "starlette>=0.47.1",
    "uvicorn>=0.34.3",
]
